Refactor code actions
