+++ metals/src/main/scala/scala/meta/internal/metals/CodeActionProvider.scala
@@ -2,27 +2,38 @@ package scala.meta.internal.metals
 
 import org.eclipse.{lsp4j => l}
 import scala.meta.pc.CancelToken
+import scala.meta.internal.metals.codeactions._
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
+import scala.collection.JavaConverters._
 
 final class CodeActionProvider(
     compilers: Compilers
 ) {
+
+  val allActions: List[CodeAction] = List(
+    new ImportMissingSymbol(compilers)
+  )
 
   def codeActions(
       params: l.CodeActionParams,
       token: CancelToken
   )(implicit ec: ExecutionContext): Future[Seq[l.CodeAction]] = {
 
-    for {
-      quickFixes <- Future
-        .sequence(
-          List(QuickFix.ImportMissingSymbol)
-            .map(_.contribute(params, compilers, token))
-        )
-        .map(_.flatten)
-    } yield quickFixes
+    def isRequestedKind(action: CodeAction): Boolean =
+      Option(params.getContext.getOnly) match {
+        case Some(only) =>
+          only.asScala.toSet.exists(requestedKind =>
+            action.kind.startsWith(requestedKind)
+          )
+        case None => true
+      }
 
+    val actions = allActions.collect {
+      case action if isRequestedKind(action) => action.contribute(params, token)
+    }
+
+    Future.sequence(actions).map(_.flatten)
   }
 
 }
