+++ metals/src/main/scala/scala/meta/internal/metals/codeactions/ImportMissingSymbol.scala
@@ -1,0 +1,63 @@
+package scala.meta.internal.metals.codeactions
+
+import scala.concurrent.Future
+import scala.meta.pc.CancelToken
+import org.eclipse.{lsp4j => l}
+import scala.concurrent.ExecutionContext
+import scala.meta.internal.metals._
+import scala.meta.internal.metals.MetalsEnrichments._
+
+class ImportMissingSymbol(compilers: Compilers) extends CodeAction {
+
+  override def kind: String = l.CodeActionKind.QuickFix
+
+  override def contribute(
+      params: l.CodeActionParams,
+      token: CancelToken
+  )(implicit ec: ExecutionContext): Future[Seq[l.CodeAction]] = {
+
+    def importMissingSymbol(
+        diagnostic: l.Diagnostic,
+        name: String
+    ): Future[Seq[l.CodeAction]] = {
+      val textDocumentPositionParams = new l.TextDocumentPositionParams(
+        params.getTextDocument(),
+        diagnostic.getRange.getEnd()
+      )
+      compilers
+        .autoImports(textDocumentPositionParams, name, token)
+        .map { imports =>
+          imports.asScala.map { i =>
+            val uri = params.getTextDocument().getUri()
+            val edit = new l.WorkspaceEdit(Map(uri -> i.edits).asJava)
+
+            val codeAction = new l.CodeAction()
+
+            codeAction.setTitle(ImportMissingSymbol.title(name, i.packageName))
+            codeAction.setKind(l.CodeActionKind.QuickFix)
+            codeAction.setDiagnostics(List(diagnostic).asJava)
+            codeAction.setEdit(edit)
+
+            codeAction
+          }
+        }
+    }
+
+    Future
+      .sequence(params.getContext().getDiagnostics().asScala.collect {
+        case d @ ScalacDiagnostic.SymbolNotFound(name)
+            if d.getRange().encloses(params.getRange().getEnd()) =>
+          importMissingSymbol(d, name)
+      })
+      .map(_.flatten)
+
+  }
+
+}
+
+object ImportMissingSymbol {
+
+  def title(name: String, packageName: String): String =
+    s"Import '$name' from package '$packageName'"
+
+}
